refactor: move student db logic to mvc
